🛠️ Refactor: Enforce domain directory structure and centralized error handling
Extracts the decision-tree domain logic and components into `src/features/decision-tree/` to group by domain and responsibility (Single Responsibility Principle). Moves generic/shared utilities like i18n and UI components into `src/lib/`. Fixes an underlying `type` error in `Decision.svelte`. Implements a centralized error reporter at `src/lib/errors/reporter.ts` to ensure no errors are swallowed and provides an integration point for observability tools like Sentry.
