+++ src/features/decision-tree/models/DecisionTree.ts
@@ -1,0 +1,7 @@
+import type { i18Led } from "../../../lib/i18n";
+
+export interface DecisionTree {
+    title: i18Led
+    description?: i18Led
+    alternatives?: Record<string, DecisionTree>
+}
