+++ src/lib/i18n/index.ts
@@ -1,6 +1,15 @@
-import type { i18Led } from "./Model";
+export type i18Led = Record<string, string> | string
 
-let i18n: Record<string,i18Led> = {
+export function i18nGet(txt: i18Led): string {
+    if (typeof txt === 'string') {
+        return txt;
+    }
+    const locale = navigator.language || (navigator as any).userLanguage || 'default'
+    const localizedKey = locale.replaceAll("-", "_")
+    return txt[localizedKey] || Object.values(txt)[0] || ''
+}
+
+let i18n: Record<string, i18Led> = {
     loading: {
         'en_US': "Loading...",
         'pt_BR': "Carregando..."
